backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
python-multipart
pandas
pytesseract
Pillow
PyMuPDF
numpy
requests
